requirements.txt
requests
tqdm
taurex
astropy
orchestral-ai
corner-2.2.3